PC04 GenArt assignment
For this assignment I created a picture of clouds and flowers. The clouds are randomly placed in the sky within specific x and y values, while the stems of the flowers are randomly placed between 5 coordinates. The actual flowers are also randomly placed between 5 different coordinates, and the colors of the flowers are random between 6 different colors. Sometimes the stems trace over each other, making one less full flower and a floating floating flower.